Add files via upload
Implementation of Neural Network using the merged_scored_dataset1. This dataset is a combination of score column which was implemented as part of clustering task to represent the district column to a numeric value.  Further, applied  Fast Gradient Method (FGM) and Basic Iterative Method (BIM)  attacks on the neural network model.